+++ src/passport_handler.erl
@@ -94,7 +94,7 @@ refreshtoken(Req0) ->
             % lager:warning("Auth ~p exceeded api limit~n", [Refreshtoken]),
             cowboy_req:reply(429, Req0);
     _ ->
-        case token_ds:decrypt_token(Refreshtoken, <<"rtk">>) of
+        case token_ds:decrypt_token(Refreshtoken) of
             {ok, Id, _ExpireAt, <<"rtk">>} ->
                 Data = [{<<"token">>, token_ds:encrypt_token(Id)}],
                 imboy_response:success(Req0, Data, "success.");